refactor: organize view/auth controllers and fix missing props in login render
