+++ controllers/authcontroller.js
@@ -1,29 +1,6 @@
 const db = require('../db/queries');
 const { validationResult } = require('express-validator');
 
-// for "/" route
-function renderRoot(req, res) {
-    if(!req.session.user) {
-        return res.render('login', {
-            csrfToken: req.csrfToken()
-        });
-    }
-    return res.render('dashboard', {
-        user: req.session.user,
-        csrfToken: req.csrfToken()
-    });
-}
-
-// for "/login" route
-// This function renders the login page
-function renderLogin(req, res) {
-    return res.render('login', {
-        csrfToken: req.csrfToken(),
-        errors: [],
-        old: {}
-    });
-}
-
 // This function checks if the user is authenticated
 // middle ware for routes that require authentication
 function ensureAuth(req, res, next) {
@@ -31,14 +8,6 @@ function ensureAuth(req, res, next) {
         return res.redirect("/login");
     }
     next();
-}
-
-// This function renders the dashboard page
-function renderDashboard(req, res) {
-    return res.render('dashboard', {
-        user: req.session.user,
-        csrfToken: req.csrfToken()
-    });
 }
 
 // This function handles the login process
@@ -71,16 +40,7 @@ async function handleLogout(req, res) {
     });
 }
 
-function renderRegister(req, res) {
-    res.render('register', {
-        csrfToken: req.csrfToken(),
-        errors: [],
-        old: {}
-    });
-}
-
 async function handleRegister(req, res) {
-
     const errors = validationResult(req);
     if (!errors.isEmpty()) {
         return res.status(422).render('register', {
@@ -101,22 +61,9 @@ async function handleRegister(req, res) {
     }
 }
 
-function renderTerms(req, res) {
-    res.render('terms', {
-        csrfToken: req.csrfToken(),
-        errors: [],
-        old: {}
-    });
-}
-
 module.exports = {
-    renderRoot,
-    renderLogin,
     ensureAuth,
-    renderDashboard,
     handleLogin,
     handleLogout,
-    renderRegister,
     handleRegister,
-    renderTerms
 };
